Convert NCzarr meta-data to use only Zarr attributes
As discussed in a netcdf meeting, convert NCZarr V2 to store all netcdf-4 specific info as attributes. This improves interoperability with other Zarr implementations by no longer using non-standard keys.

## Other Changes
* Remove support for older NCZarr formats.
* Update anonymous dimension naming
* Begin the process of fixing the -Wconversion and -Wsign-compare warnings in libnczarr, nczarr_test, and v3_nczarr_test.
* Update docs/nczarr.md
* Rebuild using the .y and .l files

+++ libnczarr/zclose.c
@@ -72,7 +72,7 @@ zclose_group(NC_GRP_INFO_T *grp)
 {
     int stat = NC_NOERR;
     NCZ_GRP_INFO_T* zgrp;
-    int i;
+    size_t i;
 
     assert(grp && grp->format_grp_info != NULL);
     LOG((3, "%s: grp->name %s", __func__, grp->hdr.name));
@@ -123,7 +123,7 @@ zclose_gatts(NC_GRP_INFO_T* grp)
 {
     int stat = NC_NOERR;
     NC_ATT_INFO_T *att;
-    int a;
+    size_t a;
     for(a = 0; a < ncindexsize(grp->att); a++) {
         NCZ_ATT_INFO_T* zatt = NULL;
         att = (NC_ATT_INFO_T* )ncindexith(grp->att, a);
@@ -149,7 +149,7 @@ NCZ_zclose_var1(NC_VAR_INFO_T* var)
     int stat = NC_NOERR;
     NCZ_VAR_INFO_T* zvar;
     NC_ATT_INFO_T* att;
-    int a;
+    size_t a;
 
     assert(var && var->format_var_info);
     zvar = var->format_var_info;;
@@ -191,7 +191,7 @@ zclose_vars(NC_GRP_INFO_T* grp)
 {
     int stat = NC_NOERR;
     NC_VAR_INFO_T* var;
-    int i;
+    size_t i;
 
     for(i = 0; i < ncindexsize(grp->vars); i++) {
         var = (NC_VAR_INFO_T*)ncindexith(grp->vars, i);
@@ -215,7 +215,7 @@ zclose_dims(NC_GRP_INFO_T* grp)
 {
     int stat = NC_NOERR;
     NC_DIM_INFO_T* dim;
-    int i;
+    size_t i;
 
     for(i = 0; i < ncindexsize(grp->dim); i++) {
         NCZ_DIM_INFO_T* zdim;
@@ -265,7 +265,7 @@ static int
 zclose_types(NC_GRP_INFO_T* grp)
 {
     int stat = NC_NOERR;
-    int i;
+    size_t i;
     NC_TYPE_INFO_T* type;
 
     for(i = 0; i < ncindexsize(grp->type); i++)
@@ -289,7 +289,7 @@ static int
 zwrite_vars(NC_GRP_INFO_T *grp)
 {
     int stat = NC_NOERR;
-    int i;
+    size_t i;
 
     assert(grp && grp->format_grp_info != NULL);
     LOG((3, "%s: grp->name %s", __func__, grp->hdr.name));
